+++ Source/TimeWarp.Blazor.Template/content/TimeWarp.Blazor/Source/Client/Components/App.razor
@@ -1,3 +1,4 @@
+@namespace TimeWarp.Blazor.Components
 @inherits AppBase
 <Router AppAssembly="typeof(Client.Program).Assembly">
   <Found Context="routeData">
