+++ Source/TimeWarp.Blazor.Template/content/TimeWarp.Blazor/Source/Server/Pages/_Host.cshtml
@@ -1,6 +1,7 @@
 ﻿@page "/"
 @namespace TimeWarp.Blazor.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@using TimeWarp.Blazor.Components
 
 <!DOCTYPE html>
 <html>
